in shambles

+++ tensor/src/profile_normal_test.cpp
@@ -1,13 +1,13 @@
 #include <chrono>
 
+#include <omp.h>
 #include "tensor.hpp"
-
 using namespace tensor;
 
 int main()
 {
-    Matrix<long> A(1024, 1024, 0);
-    Matrix<long> B(1024, 1024, 0);
+    Matrix<int> A(1024, 1024, 0);
+    Matrix<int> B(1024, 1024, 0);
     for (size_t i = 0; i < A.len_; i++)
     {
         A.matrix_[i] = i + 1;
